requirements.txt
# Core dependencies
torch>=2.0.0
torch-geometric>=2.3.0
numpy<2.0
pandas
scikit-learn
matplotlib
seaborn
tqdm
pyyaml
scipy

# Optional for better performance
networkx
